Changed activity to basic activity, created ListView
