+++ app/src/main/res/layout/content_splash_screen.xml
@@ -5,16 +5,27 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context="com.juniata.ifpizza.todoapp.SplashScreenActivity"
-    tools:showIn="@layout/app_bar_splash_screen">
+    tools:context="com.juniata.ifpizza.todoapp.SplashScreen"
+    tools:showIn="@layout/activity_splash_screen">
 
-    <TextView
-        android:layout_width="wrap_content"
+    <RelativeLayout
         android:layout_height="wrap_content"
-        android:text="Hello World!"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintLeft_toLeftOf="parent"
-        app:layout_constraintRight_toRightOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
+        android:layout_width="wrap_content"
+        xmlns:android="http://schemas.android.com/apk/res/android"
+        xmlns:tools="http://schemas.android.com/tools"
+        tools:context="com.juniata.ifpizza.todoapp.SplashScreen">
+
+        <ListView
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:id="@+id/listsList" />
+
+        <TextView
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:id="@android:id/empty"
+            android:text="@string/lists_empty"/>
+
+    </RelativeLayout>
 
 </android.support.constraint.ConstraintLayout>
