+++ app/src/main/res/values/strings.xml
@@ -1,8 +1,12 @@
 <resources>
-    <string name="app_name">ToDoApp</string>
+    <string name="app_name">ToDo</string>
 
     <string name="navigation_drawer_open">Open navigation drawer</string>
     <string name="navigation_drawer_close">Close navigation drawer</string>
 
     <string name="action_settings">Settings</string>
+    <string name="action_delete">Delete</string>
+    <string name="title_activity_splash_screen">Select a List</string>
+
+    <string name="lists_empty">No Lists Found</string>
 </resources>
